# Conflicts:
#	init-gateway.sh fixed

+++ init-gateway.sh
@@ -17,11 +17,11 @@ if [ -z "$API_SECRET" ]; then
   echo "  API_SECRET环境变量未设置， 使用默认值: $API_S"
 fi
 
-if [ -z "$BASE_DOMAIN" ]; then
+#if [ -z "$BASE_DOMAIN" ]; then
   #error '请设置BASE_DOMAIN环境变量' 1
-  BASE_DOMAIN="xpaas.net"
-  echo "  BASE_DOMAIN环境变量未设置， 使用默认值: $BASE_DOMAIN"
-fi
+#  BASE_DOMAIN="xpaas.net"
+#  echo "  BASE_DOMAIN环境变量未设置， 使用默认值: $BASE_DOMAIN"
+#fi
 
 if [ -z "$APP_SRV_NAME" ]; then
   APP_SRV_NAME="wex5"
@@ -58,49 +58,54 @@ wgstart=`expr \`date +%s%N\` / 1000000`
 echo "  等待Gateway耗时$[ wgstart - gstart ]毫秒. "
 
 outputf='  %{http_code}, 耗时: %{time_total}s. '
-#uaa
-echo -n "  /uaa 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf"  -X DELETE --url http://gateway:8001/apis/uaa
-curl -sf -o /dev/null -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=uaa" --data "upstream_url=http://uaa.$CMN_STACK_NAME:8771" --data "request_path=/uaa" --data "preserve_host=true"
-#curl -sf -o /dev/null  -w "%{http_code}
 
-#login
-echo
-echo -n "  /login 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/login
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=login" --data "upstream_url=http://uaa.$CMN_STACK_NAME:8771/uaa/login" --data "request_path=/login" --data "preserve_host=true"
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/login/plugins/ --data "name=authentication" --data "config.app_key=$API_KEY" --data "config.app_secret=$API_SECRET"
-
-#logout
-echo
-echo -n "  /logout 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/logout
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=logout" --data "upstream_url=http://uaa.$CMN_STACK_NAME:8771/uaa/logout" --data "request_path=/logout" --data "preserve_host=true"
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/logout/plugins/ --data "name=authentication" --data "config.app_key=$API_KEY" --data "config.app_secret=$API_SECRET"
-
-#db-admin
-echo
-echo -n "  /db-admin 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/db-admin
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=db-admin" --data "upstream_url=http://$APP_SRV_NAME" --data "request_path=/db-admin" --data "preserve_host=true"
-
-#storage
-echo
-echo -n "  /storage 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/storage
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=storage" --data "upstream_url=http://storage.$CMN_STACK_NAME:8770" --data "request_path=/storage" --data "preserve_host=true"
-
-#appkey
-echo
-echo -n "  /appkey 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/appkey
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=appkey" --data "upstream_url=http://appkey.$CMN_STACK_NAME:8781" --data "request_path=/appkey" --data "preserve_host=true"
-
-#sms
-echo
-echo "  sms 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/sms
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=sms" --data "upstream_url=http://sms.$CMN_STACK_NAME:8782" --data "request_path=/sms" --data "preserve_host=true"
+if [[ -z "$NO_CMNSRV" || "$NO_CMNSRV" = "false" ]]; then
+  #uaa
+  echo -n "  /uaa 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf"  -X DELETE --url http://gateway:8001/apis/uaa
+  curl -sf -o /dev/null -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=uaa" --data "upstream_url=http://uaa.$CMN_STACK_NAME:8771" --data "request_path=/uaa" --data "preserve_host=true"
+  #curl -sf -o /dev/null  -w "%{http_code}
+  
+  #login
+  echo
+  echo -n "  /login 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/login
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=login" --data "upstream_url=http://uaa.$CMN_STACK_NAME:8771/uaa/login" --data "request_path=/login" --data "preserve_host=true"
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/login/plugins/ --data "name=authentication" --data "config.app_key=$API_KEY" --data "config.app_secret=$API_SECRET"
+  
+  #logout
+  echo
+  echo -n "  /logout 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/logout
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=logout" --data "upstream_url=http://uaa.$CMN_STACK_NAME:8771/uaa/logout" --data "request_path=/logout" --data "preserve_host=true"
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/logout/plugins/ --data "name=authentication" --data "config.app_key=$API_KEY" --data "config.app_secret=$API_SECRET"
+  
+  #db-admin
+  echo
+  echo -n "  /db-admin 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/db-admin
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=db-admin" --data "upstream_url=http://$APP_SRV_NAME" --data "request_path=/db-admin" --data "preserve_host=true"
+  
+  #storage
+  echo
+  echo -n "  /storage 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/storage
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=storage" --data "upstream_url=http://storage.$CMN_STACK_NAME:8770" --data "request_path=/storage" --data "preserve_host=true"
+  
+  #appkey
+  echo
+  echo -n "  /appkey 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/appkey
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=appkey" --data "upstream_url=http://appkey.$CMN_STACK_NAME:8781" --data "request_path=/appkey" --data "preserve_host=true"
+  
+  #sms
+  echo
+  echo -n "  /sms 清理并接入 ... "
+  curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/sms
+  curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=sms" --data "upstream_url=http://sms.$CMN_STACK_NAME:8782" --data "request_path=/sms" --data "preserve_host=true"
+else
+  echo -n "  no_cmnsrv存在， 跳过公共服务接入."
+fi  
 
 #postgrest
 echo
@@ -109,29 +114,30 @@ curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/po
 curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=postgrest" --data "upstream_url=http://$APP_SRV_NAME:3000" --data "request_path=/postgrest" --data "preserve_host=true" --data "strip_request_path=true"
 
 #x5
-echo
-echo -n "  /x5 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/x5
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=x5" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/x5" --data "preserve_host=true"
+#echo
+#echo -n "  /x5 清理并接入 ... "
+#curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/x5
+#curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=x5" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/x5" --data "preserve_host=true"
 
 #ide
-echo
-echo -n "  /ide 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/ide
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=ide" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/ide" --data "preserve_host=true"
+#echo
+#echo -n "  /ide 清理并接入 ... "
+#curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/ide
+#curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=ide" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/ide" --data "preserve_host=true"
 
 #baas
-echo
-echo "  baas 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/baas
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=baas" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/baas" --data "preserve_host=true"
+#echo
+#echo "  baas 清理并接入 ... "
+#curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/baas
+#curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=baas" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/baas" --data "preserve_host=true"
 
 #/ static resource
 echo
-echo "  静态资源 清理并接入 ... "
-curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/static
-curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=static" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/" --data "preserve_host=true"
+echo -n "  应用 清理并接入 ... "
+curl -sf -o /dev/null  -w "$outputf" -X DELETE --url http://gateway:8001/apis/app
+curl -sf -o /dev/null  -w "$outputf" -X POST --url http://gateway:8001/apis/ --data "name=app" --data "upstream_url=http://$APP_SRV_NAME:$APP_SRV_PORT" --data "request_path=/" --data "preserve_host=true"
 
 execute=`expr \`date +%s%N\` / 1000000`
 echo
 echo "  网关初始化调用完成. 耗时$[ execute - wgstart ]毫秒."
+
